requirements.txt
flask>=2.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.1
waitress>=3.0.0
